Make sure we delete the emails from POP3 after we have downloaded them.
This is not ideal, if we fail during processing of the first message we
will lose all of the following emails.  This is now great.

Really we should fetch and persist the email, and then work on the
persisted version before deleting *after* we're sure the response email
has been sent.